pyproject.toml
[tool.poetry]
name = "urbanopt-des"
version = "0.1.0"
description = ""
authors = ["Nicholas Long <nicholas.long@nrel.gov>"]
readme = "README.md"
packages = [{ include = "urbanopt_des", from = "." }]

[tool.poetry.dependencies]
python = ">=3.10,<3.14"
geopandas = "^1.0.1"
cyclopts = "^3.10.1"
# release mode
geojson-modelica-translator = "^0.9.3"
# pre-release mode, use github
# geojson-modelica-translator = { git = "https://github.com/urbanopt/geojson-modelica-translator.git", branch = "develop" }
# dev mode
# geojson-modelica-translator = { path = "../geojson-modelica-translator", develop = true }

[tool.poetry.group.dev.dependencies]
pre-commit = "~=4.0"
mypy = "~1.14"
pytest = "~=8.3"
pytest-cov = "~6.0"
buildingspy = "^5.2.0"

[build-system]
requires = ["poetry-core>=2.0.1"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
des-output = "urbanopt_des.cli:app"
